+++ cpp/src/error.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024-2025, NVIDIA CORPORATION.
+ * Copyright (c) 2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,19 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#pragma once
 
-#include <kvikio/utils.hpp>
-
-namespace kvikio {
-
-/**
- * @brief Get the filepath to cuFile's config file (`cufile.json`) or the empty string
- *
- * This lookup is cached.
- *
- * @return The filepath to the cufile.json file or the empty string if it isn't found.
- */
-[[nodiscard]] KVIKIO_EXPORT const std::string& config_path();
-
-}  // namespace kvikio
+#include <kvikio/error.hpp>
